Just found all the documentation on the github wiki thats not in the sphinx documentation..  renaming amimodel to ami_model is too breaking to all of that documentation as well.
