+++ pyibisami/amimodel.py
